[CompositeClipPathAnimations] Fix a crash when a keyframe in an animation has clip-path: none
When composited clip path animations are enabled, clip path animations
that have a keyframe with clip-path: none crash.

This change adds a filter function to GetAnimationIfCompositable to
exclude animations with clip-path: none.

A clip path of none is difficult to paint with composited clip path as currently implemented, and there is no great need to composite this animation given there is no continuous interpolation to be done.

Bug: 1345977
Change-Id: I2bed8ab0ed6a41b7341f92a3bd48cc2e6457fcb4