[grid] Small fixes & remove complex baseline logic.
Previously we had complicated "truth tables" to determine the correct
baseline. These were un-intuitive, and complex.

This patch introduces the concept of a writing-mode to read a baseline
for a particular grid-item.

This is what the spec(s) use, and should be easier to keep consistent
with the spec.

The two tests updated were using the wrong edge to synthesize their
respective baselines.

Change-Id: I576b79444e27e03583c709eb15e9e13d400e8b94