Deflake pointerevent_boundary_events_in_capturing.html
The test was expecting at least two pointermove events from a drag
gesture.  When MT is slow, those pointermove events can be coalesced
into a single one.  Rewrote the test to remove the assumption, now
either a pointermove or a pointerup completes the required checks.

Also switched the test to use promise_test and a simpler action/event
sequencing for better readablity.

Fixed: 1289232
Change-Id: I3a0dcb52a814fd21cce54e8a5ab6af3ed077be5b